# Enable guava-beta-checker and fix findings