--- a/infrastructure/bls/src/main/java/tech/pegasys/teku/bls/impl/PublicKeyMessagePair.java
+++ b/infrastructure/bls/src/main/java/tech/pegasys/teku/bls/impl/PublicKeyMessagePair.java
@@ -15,9 +15,8 @@ package tech.pegasys.teku.bls.impl;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
-import com.google.common.collect.Streams;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 import org.apache.tuweni.bytes.Bytes;
 
 /**
@@ -30,8 +29,11 @@ public class PublicKeyMessagePair {
   public static List<PublicKeyMessagePair> fromLists(
       List<PublicKey> publicKeys, List<Bytes> messages) {
     checkArgument(publicKeys.size() == messages.size());
-    return Streams.zip(publicKeys.stream(), messages.stream(), PublicKeyMessagePair::new)
-        .collect(Collectors.toList());
+    List<PublicKeyMessagePair> publicKeyMessagePairs = new ArrayList<>();
+    for (int i = 0; i < publicKeys.size(); i++) {
+      publicKeyMessagePairs.add(new PublicKeyMessagePair(publicKeys.get(i), messages.get(i)));
+    }
+    return publicKeyMessagePairs;
   }
 
   private final PublicKey publicKey;