## PR Description

[`guava-beta-checker`](https://github.com/google/guava-beta-checker) is Google's EP checker that identifies usage of beta Guava APIs. This checker is limited to the main source tree and doesn't produce warnings for test code. There weren't many findings.

> Such APIs should never be used in library code that other projects may depend on; using the Beta Checker can help library projects ensure that they don't use them.

When removing `Streams.zip` I was worried that the new implementation might be slower, so I added a benchmark function for the `BLS.aggregateVerify` method. In the process of adding this, I made a some changes to whole benchmark class; things could be simpler with a setup function.

#### Before

```
Benchmark                              (sigCnt)   Mode  Cnt    Score   Error  Units
BLSBenchmark.aggregateVerifySignature         1  thrpt   10  926.271 ± 1.448  ops/s
BLSBenchmark.aggregateVerifySignature         2  thrpt   10  710.323 ± 1.462  ops/s
BLSBenchmark.aggregateVerifySignature         4  thrpt   10  481.074 ± 5.471  ops/s
BLSBenchmark.aggregateVerifySignature         8  thrpt   10  295.479 ± 2.709  ops/s
BLSBenchmark.aggregateVerifySignature        16  thrpt   10  163.657 ± 1.735  ops/s
BLSBenchmark.aggregateVerifySignature        32  thrpt   10   86.401 ± 0.738  ops/s
BLSBenchmark.aggregateVerifySignature        64  thrpt   10   44.565 ± 0.155  ops/s
BLSBenchmark.aggregateVerifySignature       128  thrpt   10   22.638 ± 0.093  ops/s
```

#### After

```
Benchmark                              (sigCnt)   Mode  Cnt    Score   Error  Units
BLSBenchmark.aggregateVerifySignature         1  thrpt   10  927.673 ± 1.370  ops/s
BLSBenchmark.aggregateVerifySignature         2  thrpt   10  712.032 ± 0.952  ops/s
BLSBenchmark.aggregateVerifySignature         4  thrpt   10  485.891 ± 2.164  ops/s
BLSBenchmark.aggregateVerifySignature         8  thrpt   10  294.099 ± 1.980  ops/s
BLSBenchmark.aggregateVerifySignature        16  thrpt   10  162.698 ± 1.052  ops/s
BLSBenchmark.aggregateVerifySignature        32  thrpt   10   86.303 ± 0.847  ops/s
BLSBenchmark.aggregateVerifySignature        64  thrpt   10   44.676 ± 0.309  ops/s
BLSBenchmark.aggregateVerifySignature       128  thrpt   10   22.720 ± 0.018  ops/s
```

So things are pretty much the exact same. The new version might actually be a hair faster.

## Documentation

- [x] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [x] I thought about adding a changelog entry, and added one if I deemed necessary.