@jtraglia 2022-06-15
On second thought, Teku isn't a library and this restriction doesn't really apply; we control the version of Guava we use. If a beta API is modified, we will know. So unless this is something you guys want, I'll close this PR.