@rolfyone 2022-06-15
The Streams.zip is fairly neat, so we might stick with the current implementation :)