@jtraglia 2022-06-15
Yup I agree. Sorry to waste your time. Thanks!