@rolfyone 2022-06-16
I'd never say anything like this is a waste of time :) thanks for your continued efforts!